requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.3
idna==3.4
importlib-resources==5.12.0
kiwisolver==1.4.4
logzero==1.7.0
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.29.0
six==1.16.0
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.15
zipp==3.15.0
